requirements.txt
Django==4.1.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
stockfish
